kernel/debug : Move function of saving termination info for debug into task_exithook
For stack monitor and heapinfo, we should save the information of terminated thread.
In previous, sometimes tcb->stack_alloc_ptr can be NULL, because of group_leave.
So move the function into task_exithook which is calling before group_leave.
